Properly check dotenv keys
Rename utils_cli -> cli

Remove all CLI commands except 'config' and 'upload'.
