+++ dlite_entities_service/cli/__init__.py
