+++ dlite_entities_service/cli/config.py
@@ -29,8 +29,8 @@ from dotenv import dotenv_values, set_key, unset_key
 from rich import print, print_json  # pylint: disable=redefined-builtin
 from rich.console import Console
 
+from dlite_entities_service.cli._utils.global_settings import STATUS
 from dlite_entities_service.service.config import CONFIG
-from dlite_entities_service.utils_cli._utils.global_settings import STATUS
 
 ERROR_CONSOLE = Console(stderr=True)
 CLI_DOTENV_FILE: Path = (
